implemented single routes

+++ src/components/Blog/Blog.tsx
@@ -1,13 +1,16 @@
 import { useRef, useState } from 'react';
+import authService from '../../services/auth.service';
 import { BlogAttributes, BlogUpdate } from '../../types/blog.type';
+import { dateToString } from '../../util';
 import Card from '../Card/Card';
 import Editable, { EditableRef } from '../Editable/Editable';
-import TextLink from '../TextLink/TextLink';
 import { BlogAuthor, BlogBody, BlogLikes, BlogLinkContainer } from './Blog.styled';
+import InternalLink from '../Link/InternalLink';
+import ExternalLink from '../Link/ExternalLink';
 
 export interface BlogCallbacks {
-  onUpdate: (blog: BlogAttributes) => void;
-  onDelete: (blog: BlogAttributes) => void;
+  onSave?: (blog: BlogAttributes) => void;
+  onDelete?: (blog: BlogAttributes) => void;
 }
 
 interface Props extends BlogCallbacks {
@@ -18,9 +21,14 @@ export interface BlogInnerProps {
   warning?: boolean;
 }
 
-const Blog = ({ blog, onUpdate, onDelete }: Props) => {
+const Blog = ({ blog, onSave, onDelete }: Props) => {
   const [editable, setEditable] = useState(false);
   const [warning, setWarning] = useState(false);
+  const auth = blog.owner?.username === authService.getUser()?.username;
+
+  console.log(!!(auth && (onSave || onDelete)), auth, onSave, onDelete);
+
+  const enableEdit = !!(auth && (onSave || onDelete));
 
   const tabIndex = { tabIndex: warning ? -1 : 0 };
 
@@ -45,18 +53,23 @@ const Blog = ({ blog, onUpdate, onDelete }: Props) => {
 
     if (Object.keys(update).length > 0) {
       const data: BlogAttributes = { ...blog, ...update };
-      onUpdate(data);
+      onSave && onSave(data);
     }
   };
 
   const deleteHandler = () => {
-    onDelete(blog);
+    onDelete && onDelete(blog);
   };
+
+  if (!blog) return null;
+
+  const createdAt = dateToString(blog.createdAt);
 
   return (
     <Card
-      onSave={saveHandler}
-      onDelete={deleteHandler}
+      enableEdit={enableEdit}
+      onSave={onSave && saveHandler}
+      onDelete={onDelete && deleteHandler}
       onEdit={setEditable}
       onWarning={setWarning}
       header={<Editable tagName="h2" ref={title} initialValue={blog.title} disabled={!editable} />}
@@ -66,21 +79,37 @@ const Blog = ({ blog, onUpdate, onDelete }: Props) => {
           by &nbsp;
           <Editable tagName="span" ref={author} initialValue={blog.author} disabled={!editable} />
         </BlogAuthor>
-
         <BlogLikes>has {blog.likes} likes</BlogLikes>
-
         {editable ? (
           <Editable tagName="span" ref={url} initialValue={blog.url} disabled={!editable} />
         ) : (
           <BlogLinkContainer>
-            <TextLink href={blog.url} target="_blank" truncate {...tabIndex}>
+            <ExternalLink href={blog.url} truncate {...tabIndex}>
               {blog.url}
-            </TextLink>
+            </ExternalLink>
           </BlogLinkContainer>
         )}
+        <div>{createdAt && <div>created: {createdAt}</div>}</div>
+        <div>Owner?: {blog.owner && <InternalLink to={`/users/${blog.owner.id}`}>{blog.owner.name}</InternalLink>}</div>
+        <Readers readers={blog.readers} />
       </BlogBody>
     </Card>
   );
+};
+
+type ReadersProps = Pick<BlogAttributes, 'readers'>;
+
+const Readers = ({ readers }: ReadersProps) => {
+  if (!readers || readers.length === 0) return 'No Readers';
+  return readers.map((reader, i) => (
+    <div key={i}>
+      Readers:{' '}
+      <div>
+        <div>Name:{reader.name}</div>
+        <div>Read:{reader.reading.read ? 'Yes' : 'No'}</div>
+      </div>
+    </div>
+  ));
 };
 
 export default Blog;
