+++ src/components/User/User.tsx
@@ -4,15 +4,15 @@ import Card from '../Card/Card';
 import Editable, { EditableRef } from '../Editable/Editable';
 
 export interface UserCallbacks {
-  onUpdate: (user: UserAttributes) => void;
-  onDelete: (user: UserAttributes) => void;
+  onSave?: (user: UserAttributes) => void;
+  onDelete?: (user: UserAttributes) => void;
 }
 
 interface Props extends UserCallbacks {
   user: UserAttributes;
 }
 
-const User = ({ user, onUpdate, onDelete }: Props) => {
+const User = ({ user, onSave, onDelete }: Props) => {
   const [editable, setEditable] = useState(false);
 
   const name = useRef<EditableRef>(null);
@@ -26,22 +26,52 @@ const User = ({ user, onUpdate, onDelete }: Props) => {
 
     if (Object.keys(update).length > 0) {
       const data: UserAttributes = { ...user, ...update };
-      onUpdate(data);
+      onSave && onSave(data);
     }
   };
 
   const deleteHandler = () => {
-    onDelete(user);
+    onDelete && onDelete(user);
   };
+
+  if (!user) return null;
 
   return (
     <Card
-      onSave={saveHandler}
-      onDelete={deleteHandler}
+      onSave={onSave && saveHandler}
+      onDelete={onDelete && deleteHandler}
       onEdit={setEditable}
       header={<Editable tagName="h2" ref={name} initialValue={user.name} disabled={!editable} />}
     >
       {user.username}
+      {user.disabled && <div>disabled!</div>}
+      <div>Privileges: {user.admin ? 'Admin' : 'User'}</div>
+
+      <div>
+        {user.blogs && user.blogs.length > 0 && (
+          <>
+            <h3>Blogs</h3>
+            <ul>
+              {user.blogs.map((blog) => (
+                <li key={blog.id}>{blog.title}</li>
+              ))}
+            </ul>
+          </>
+        )}
+      </div>
+
+      <div>
+        {user.readings && user.readings.length > 0 && (
+          <>
+            <h3>Readings</h3>
+            <ul>
+              {user.readings.map((reading) => (
+                <li key={reading.id}>{reading.title}</li>
+              ))}
+            </ul>
+          </>
+        )}
+      </div>
     </Card>
   );
 };
